Move traits to kernel.
