+++ src/Applications/OfficialAccount/Card/Client.php
@@ -13,7 +13,7 @@ namespace EasyWeChat\Applications\OfficialAccount\Card;
 
 use EasyWeChat\Kernel\BaseClient;
 use EasyWeChat\Support\Arr;
-use EasyWeChat\Support\InteractsWithCache;
+use EasyWeChat\Kernel\Traits\InteractsWithCache;
 use Psr\Http\Message\ResponseInterface;
 
 /**
