+++ src/Kernel/Traits/InteractsWithCache.php
@@ -9,7 +9,7 @@
  * with this source code in the file LICENSE.
  */
 
-namespace EasyWeChat\Support;
+namespace EasyWeChat\Kernel\Traits;
 
 use Psr\SimpleCache\CacheInterface;
 use Symfony\Component\Cache\Simple\FilesystemCache;
